+++ src/main/java/ru/otus/mkulikov/app/model/Genre.java
@@ -12,16 +12,16 @@ import lombok.Data;
 @Data
 public class Genre {
 
-    private final int id;
+    private final long id;
     private final String name;
 
-    public Genre(int id, String name) {
+    public Genre(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
     public Genre(String name) {
-        this.id = 0;
+        this.id = 0L;
         this.name = name;
     }
